+++ packages/overture-schema-foundation/src/overture/schema/foundation/primitive/bbox.py
